renaming the feature flag to 'beta_dropdown'
