+++ app/components/viral/dropdown_component.html.erb
@@ -1,9 +1,9 @@
 <div
   class="flex max-sm:grow"
-  <% if Flipper.enabled?(:flowbite_replacement) %>
-  data-controller="viral--dropdown<%= tooltip? ? ' pathogen--tooltip' : '' %>"
+  <% if Flipper.enabled?(:beta_dropdown) %>
+  data-controller="viral--beta-dropdown<%= tooltip? ? ' pathogen--tooltip' : '' %>"
   <% else %>
-  data-controller="viral--flowbite-dropdown<%= tooltip? ? ' pathogen--tooltip' : '' %>"
+  data-controller="viral--dropdown<%= tooltip? ? ' pathogen--tooltip' : '' %>"
   <% end %>
   data-viral--dropdown-trigger-type-value="none"
   data-viral--dropdown-skidding-value="<%= skidding %>"
@@ -32,12 +32,12 @@
   <!-- Dropdown menu -->
   <ul
     id="<%= @dd_id %>"
-    <% if Flipper.enabled?(:flowbite_replacement) %>
-    data-viral--dropdown-target="menu"
+    <% if Flipper.enabled?(:beta_dropdown) %>
+    data-viral--beta-dropdown-target="menu"
     <% else %>
-    data-viral--flowbite-dropdown-target="menu"
-    data-viral--flowbite-dropdown-skidding-value="<%= skidding %>"
-    data-viral--flowbite-dropdown-distance-value="<%= distance %>"
+    data-viral--dropdown-target="menu"
+    data-viral--dropdown-skidding-value="<%= skidding %>"
+    data-viral--dropdown-distance-value="<%= distance %>"
     <% end %>
     role="menu"
     class="
