+++ app/components/viral/select2_component.html.erb
@@ -1,6 +1,6 @@
 <%= tag.div(
   data: {
-    action: Flipper.enabled?(:flowbite_replacement) ? "keydown->viral--select2#keydown" : "keydown->viral--flowbite-select2#keydown",
+    action: Flipper.enabled?(:beta_dropdown) ? "keydown->viral--beta-select2#keydown" : "keydown->viral--select2#keydown",
   },
   class: "select2"
   ) do %>
@@ -9,12 +9,12 @@
     id: id,
     name: id,
     data: {
-      if Flipper.enabled?(:flowbite_replacement)
-        "viral--select2-target" => "trigger",
-        action: "input->viral--select2#input",
+      if Flipper.enabled?(:beta_dropdown)
+        "viral--beta-select2-target": "trigger",
+        action: "input->viral--beta-select2#input",
       else
-        "viral--flowbite-select2-target" => "input",
-        action: "input->viral--flowbite-select2#input",
+        "viral--select2-target": "input",
+        action: "input->viral--select2#input",
       end
     },
     placeholder: placeholder,
@@ -31,27 +31,27 @@
   ) %>
 
   <%= form.hidden_field name, data: {
-    if Flipper.enabled?(:flowbite_replacement)
-      "viral--select2-target": "hidden",
+    if Flipper.enabled?(:beta_dropdown)
+      "viral--beta-select2-target": "hidden",
     else
-      "viral--flowbite-select2-target" => "hidden",
+      "viral--select2-target": "hidden",
     end
   } %>
 
   <%= tag.div(
     data: {
-      if Flipper.enabled?(:flowbite_replacement)
-        "viral--select2-target": "menu",
+      if Flipper.enabled?(:beta_dropdown)
+        "viral--beta-select2-target": "menu",
       else
-        "viral--flowbite-select2-target" => "dropdown",
+        "viral--select2-target": "dropdown",
       end
       },
     class: "z-10 w-full bg-white divide-y rounded-lg shadow-sm divide-slate-100 dark:bg-slate-700",
     hidden:""
   ) do %>
     <%= tag.ul(
-      if Flipper.enabled?(:flowbite_replacement)
-        data: { "viral--select2-target": "scroller" },
+      if Flipper.enabled?(:beta_dropdown)
+        data: { "viral--beta-select2-target": "scroller" },
         class: "w-full py-2 overflow-y-auto divide-y divide-slate-200 dark:divide-slate-800",
         style:"max-height: inherit;"
       else
@@ -75,10 +75,10 @@
           class:
             "p-4 text-blue-800 border border-blue-300 bg-blue-50 dark:bg-slate-800 dark:text-blue-400 dark:border-blue-800 pointer-events-none",
           data: {
-            if Flipper.enabled?(:flowbite_replacement)
-              "viral--select2-target": "empty",
+            if Flipper.enabled?(:beta_dropdown)
+              "viral--beta-select2-target": "empty",
             else
-            "viral--flowbite-select2-target": "empty",
+            "viral--select2-target": "empty",
             end
           },
           hidden:""
