+++ app/javascript/controllers/viral/beta_dropdown_controller.js
@@ -1,23 +1,29 @@
-import { Controller } from "@hotwired/stimulus";
+import MenuController from "controllers/menu_controller";
 
-export default class extends Controller {
+export default class extends MenuController {
   static targets = ["trigger", "menu"];
-  static values = {
-    position: String,
-    trigger: String,
-    skidding: Number,
-    distance: Number,
-  };
 
   initialize() {
     this.boundOnButtonKeyDown = this.onButtonKeyDown.bind(this);
     this.boundOnButtonClick = this.onButtonClick.bind(this);
     this.boundOnMenuItemKeyDown = this.onMenuItemKeyDown.bind(this);
     this.boundFocusOut = this.focusOut.bind(this);
+    this.boundOnMorph = this.onMorph.bind(this);
   }
 
   connect() {
     this.element.setAttribute("data-controller-connected", "true");
+    document.addEventListener("turbo:morph", this.boundOnMorph);
+    this.#initializeDropdown();
+  }
+
+  disconnect() {
+    document.removeEventListener("turbo:morph", this.boundOnMorph);
+  }
+
+  onMorph() {
+    this.menuTargetConnected(this.menuTarget);
+    this.triggerTargetConnected(this.triggerTarget);
   }
 
   menuTargetConnected(element) {
@@ -39,29 +45,36 @@ export default class extends Controller {
     element.addEventListener("click", this.boundOnButtonClick, {
       capture: true,
     });
-    this.dropdown = new Dropdown(this.menuTarget, element, {
-      triggerType: "none",
-      offsetSkidding: this.skiddingValue,
-      offsetDistance: this.distanceValue,
-      onShow: () => {
-        this.triggerTarget.setAttribute("aria-expanded", "true");
-        this.menuTarget.setAttribute("aria-hidden", "false");
-        this.menuTarget.removeAttribute("hidden");
-      },
-      onHide: () => {
-        this.triggerTarget.setAttribute("aria-expanded", "false");
-        this.menuTarget.setAttribute("aria-hidden", "true");
-        this.menuTarget.setAttribute("hidden", "hidden");
-        this.#menuItems(element).forEach((menuitem) => {
-          menuitem.setAttribute("tabindex", "-1");
-        });
-      },
+    super.triggerTargetConnected();
+  }
+
+  #initializeDropdown() {
+    super.share({
+      onShow: () => this.#onShow(),
+      onHide: () => this.#onHide(),
     });
+    super.connect();
+  }
+
+  #onShow() {
+    this.triggerTarget.setAttribute("aria-expanded", "true");
+    this.menuTarget.setAttribute("aria-hidden", "false");
+    this.menuTarget.removeAttribute("hidden");
+  }
+
+  #onHide() {
+    this.triggerTarget.setAttribute("aria-expanded", "false");
+    this.menuTarget.setAttribute("aria-hidden", "true");
+    this.menuTarget.setAttribute("hidden", "");
+    this.#menuItems(this.triggerTarget).forEach((menuitem) => {
+      menuitem.setAttribute("tabindex", "-1");
+    });
+    this.triggerTarget.focus();
   }
 
   focusOut(event) {
     if (!this.element.contains(event.relatedTarget)) {
-      this.dropdown.hide();
+      super.hide();
     }
   }
 
@@ -69,8 +82,8 @@ export default class extends Controller {
     event.preventDefault();
     event.stopPropagation();
 
-    if (this.dropdown.isVisible()) {
-      this.dropdown.hide();
+    if (super.isVisible()) {
+      super.hide();
     } else {
       this.#openMenuAndFocusMenuItem(0);
     }
@@ -108,9 +121,10 @@ export default class extends Controller {
         },
         { once: true },
       );
-      this.dropdown.show();
+      super.show();
+      this.#focusMenuItem(this.menuTarget);
     } else {
-      this.dropdown.show();
+      super.show();
       this.#focusMenuItem(menuItems.at(index));
     }
   }
@@ -135,7 +149,7 @@ export default class extends Controller {
           if (clickableTarget) {
             clickableTarget.click();
           } else {
-            this.dropdown.hide();
+            super.hide();
           }
         } else {
           menuItems[currentIndex].click();
@@ -149,7 +163,7 @@ export default class extends Controller {
         );
       case "Escape":
         event.preventDefault();
-        this.triggerTarget.focus();
+        super.hide();
         break;
       case "ArrowUp": {
         event.preventDefault();
@@ -185,7 +199,7 @@ export default class extends Controller {
         if (event.shiftKey) {
           event.preventDefault();
           this.triggerTarget.focus();
-          this.dropdown.hide();
+          super.hide();
         }
         break;
     }
