+++ app/javascript/controllers/viral/beta_select2_controller.js
@@ -1,4 +1,4 @@
-import { Controller } from "@hotwired/stimulus";
+import MenuController from "controllers/menu_controller";
 
 /**
  * Select2Controller
@@ -10,11 +10,11 @@ import { Controller } from "@hotwired/stimulus";
  * - Dropdown positioning and focus management
  * - Submit button enable/disable logic
  */
-export default class Select2Controller extends Controller {
+export default class Select2Controller extends MenuController {
   static targets = [
-    "input",
+    "trigger",
     "hidden",
-    "dropdown",
+    "menu",
     "scroller",
     "item",
     "empty",
@@ -26,7 +26,6 @@ export default class Select2Controller extends Controller {
   #itemSelected = false;
   #cachedInputValue = "";
   #currentItemIndex = -1;
-  #dropdown = null;
   #boundHandlers = {};
 
   static #KEY_CODES = {
@@ -46,16 +45,16 @@ export default class Select2Controller extends Controller {
 
       this.#boundHandlers.dropdownFocusOut =
         this.#handleDropdownFocusOut.bind(this);
-      this.dropdownTarget.addEventListener(
+      this.menuTarget.addEventListener(
         "focusout",
         this.#boundHandlers.dropdownFocusOut,
       );
 
       // Accessibility: set ARIA attributes
-      this.inputTarget.setAttribute("role", "combobox");
-      this.inputTarget.setAttribute("aria-autocomplete", "list");
-      this.inputTarget.setAttribute("aria-expanded", "false");
-      this.inputTarget.setAttribute(
+      this.triggerTarget.setAttribute("role", "combobox");
+      this.triggerTarget.setAttribute("aria-autocomplete", "list");
+      this.triggerTarget.setAttribute("aria-expanded", "false");
+      this.triggerTarget.setAttribute(
         "aria-controls",
         this.scrollerTarget.id || "select2-listbox",
       );
@@ -76,15 +75,12 @@ export default class Select2Controller extends Controller {
   disconnect() {
     try {
       if (this.#boundHandlers.dropdownFocusOut) {
-        this.dropdownTarget.removeEventListener(
+        this.menuTarget.removeEventListener(
           "focusout",
           this.#boundHandlers.dropdownFocusOut,
         );
       }
-      if (this.#dropdown) {
-        this.#dropdown.hide();
-        this.#dropdown = null;
-      }
+      super.hide();
     } catch (error) {
       this.#handleError(error, "disconnect");
     }
@@ -119,8 +115,8 @@ export default class Select2Controller extends Controller {
       if (selectedItemData) {
         const { value, label } = selectedItemData;
         this.#updateSelection(value, label);
-        if (this.#dropdown) this.#dropdown.hide();
-        this.inputTarget.focus();
+        super.hide();
+        this.triggerTarget.focus();
       } else {
         // If no valid item was determined (edge case or unexpected state),
         // potentially reset or log, but avoid throwing an error unless critical.
@@ -162,7 +158,7 @@ export default class Select2Controller extends Controller {
           break;
         case Select2Controller.#KEY_CODES.ENTER:
           if (this.#currentItemIndex < 0) {
-            if (this.#dropdown) this.#dropdown.show();
+            super.show();
           } else {
             this.select(event);
           }
@@ -175,8 +171,7 @@ export default class Select2Controller extends Controller {
 
   showDropdown() {
     try {
-      if (this.#dropdown) this.#dropdown.show();
-      this.inputTarget.setAttribute("aria-expanded", "true");
+      super.show();
     } catch (error) {
       this.#handleError(error, "showDropdown");
     }
@@ -184,8 +179,7 @@ export default class Select2Controller extends Controller {
 
   hideDropdown() {
     try {
-      if (this.#dropdown) this.#dropdown.hide();
-      this.inputTarget.setAttribute("aria-expanded", "false");
+      super.hide();
     } catch (error) {
       this.#handleError(error, "hideDropdown");
     }
@@ -196,7 +190,7 @@ export default class Select2Controller extends Controller {
    */
   input() {
     try {
-      const query = this.inputTarget.value.toLowerCase().trim();
+      const query = this.triggerTarget.value.toLowerCase().trim();
       let visibleItemCount = 0;
 
       this.#setItemSelected(false);
@@ -204,10 +198,10 @@ export default class Select2Controller extends Controller {
       this.itemTargets.forEach((item) => {
         const text = item.textContent.toLowerCase() || "";
         if (text.includes(query)) {
-          item.removeAttribute("hidden");
+          item.classList.remove("hidden");
           visibleItemCount++;
         } else {
-          item.setAttribute("hidden", "");
+          item.classList.add("hidden");
         }
       });
 
@@ -215,8 +209,8 @@ export default class Select2Controller extends Controller {
       this.#updateAriaActiveDescendant();
 
       if (visibleItemCount > 0) {
-        if (this.#dropdown && !this.#dropdown.isVisible()) {
-          this.#dropdown.show();
+        if (!super.isVisible()) {
+          super.show();
         }
         this.emptyTarget.setAttribute("hidden", "");
         this.scrollerTarget.scrollTop = 0;
@@ -247,8 +241,8 @@ export default class Select2Controller extends Controller {
       );
       return; // Prevent updating with invalid data
     }
-    this.inputTarget.value = label;
-    this.inputTarget.title = label;
+    this.triggerTarget.value = label;
+    this.triggerTarget.title = label;
     this.#cachedInputValue = value; // Cache the *value*, not the label
     this.hiddenTarget.value = value;
     this.#updateAriaSelected(label); // Update ARIA state based on the new selection
@@ -265,17 +259,17 @@ export default class Select2Controller extends Controller {
 
   #resetInput() {
     try {
-      if (this.#dropdown) this.#dropdown.hide();
+      super.hide();
       if (this.#cachedInputValue) {
         this.hiddenTarget.value = this.#cachedInputValue;
         this.#setInputTargetValueFromCache();
         this.#setItemSelected(true);
       } else {
-        this.inputTarget.value = "";
+        this.triggerTarget.value = "";
         this.hiddenTarget.value = "";
         this.#setItemSelected(false);
       }
-      this.inputTarget.focus();
+      this.triggerTarget.focus();
       this.#updateAriaActiveDescendant();
 
       if (this.hasSpreadsheetImportOutlet) {
@@ -337,58 +331,32 @@ export default class Select2Controller extends Controller {
       // Check if the item itself or its immediate parent container is hidden
       const parentElement = item.closest("li") || item.parentNode; // Adjust selector if needed
       return (
-        !item.hasAttribute("hidden") && !parentElement.hasAttribute("hidden")
+        !item.classList.contains("hidden") &&
+        !parentElement.classList.contains("hidden")
       );
     });
   }
 
   #ensureItemVisible(item) {
     if (!this.scrollerTarget || !item) return;
-    const container = this.scrollerTarget;
-    const containerRect = container.getBoundingClientRect();
-    const itemRect = item.getBoundingClientRect();
-    if (itemRect.bottom > containerRect.bottom) {
-      container.scrollTop += itemRect.bottom - containerRect.bottom;
-    } else if (itemRect.top < containerRect.top) {
-      container.scrollTop -= containerRect.top - itemRect.top;
-    }
+    item.scrollIntoView();
   }
 
   #initializeDropdown() {
-    try {
-      if (typeof Dropdown !== "function") {
-        throw new Error(
-          "Flowbite Dropdown class not found. Make sure Flowbite JS is loaded.",
-        );
-      }
-      this.#dropdown = new Dropdown(this.dropdownTarget, this.inputTarget, {
-        placement: "bottom",
-        triggerType: "click",
-        offsetSkidding: 0,
-        offsetDistance: 10,
-        delay: 300,
-        onShow: () => {
-          this.dropdownTarget.style.minWidth = `${this.inputTarget.offsetWidth}px`;
-          this.inputTarget.setAttribute("aria-expanded", "true");
-          this.dropdownTarget.setAttribute("aria-hidden", "false");
-          this.dropdownTarget.removeAttribute("hidden");
-        },
-        onHide: () => {
-          this.inputTarget.setAttribute("aria-expanded", "false");
-          this.dropdownTarget.setAttribute("aria-hidden", "true");
-          this.dropdownTarget.setAttribute("hidden", "hidden");
-          if (!this.#itemSelected) this.#setInputTargetValueFromCache();
-        },
-      });
-    } catch (error) {
-      this.#handleError(error, "initializeDropdown");
-    }
+    super.share({
+      onHide: () => this.#onHide(),
+    });
+    super.connect();
+  }
+
+  #onHide() {
+    if (!this.#itemSelected) this.#setInputTargetValueFromCache();
   }
 
   #setDefaultSelection() {
     try {
-      if (!this.inputTarget.value) return;
-      const query = this.inputTarget.value;
+      if (!this.triggerTarget.value) return;
+      const query = this.triggerTarget.value;
       let matched = false;
       for (const item of this.itemTargets) {
         const { value, label } = item.dataset;
@@ -399,9 +367,9 @@ export default class Select2Controller extends Controller {
           break;
         }
       }
-      if (!matched && this.inputTarget.value) {
+      if (!matched && this.triggerTarget.value) {
         throw new Error(
-          "No matching item found for the input value. Please check your data.",
+          "No matching item found for the trigger value. Please check your data.",
         );
       }
     } catch (error) {
@@ -411,8 +379,8 @@ export default class Select2Controller extends Controller {
 
   #handleDropdownFocusOut(event) {
     try {
-      if (!this.dropdownTarget.contains(event.relatedTarget)) {
-        if (this.#dropdown) this.#dropdown.hide();
+      if (!this.menuTarget.contains(event.relatedTarget)) {
+        super.hide();
       }
     } catch (error) {
       this.#handleError(error, "handleDropdownFocusOut");
@@ -421,9 +389,9 @@ export default class Select2Controller extends Controller {
 
   #validateTargets() {
     const missingTargets = [];
-    if (!this.hasInputTarget) missingTargets.push("input");
+    if (!this.hasTriggerTarget) missingTargets.push("trigger");
     if (!this.hasHiddenTarget) missingTargets.push("hidden");
-    if (!this.hasDropdownTarget) missingTargets.push("dropdown");
+    if (!this.hasMenuTarget) missingTargets.push("menu");
     if (!this.hasScrollerTarget) missingTargets.push("scroller");
     if (missingTargets.length > 0) {
       throw new Error(`Missing required targets: ${missingTargets.join(", ")}`);
@@ -436,7 +404,7 @@ export default class Select2Controller extends Controller {
   }
 
   #setInputTargetValueFromCache() {
-    const inputValue = this.inputTarget.value;
+    const inputValue = this.triggerTarget.value;
     if (inputValue === this.#cachedInputValue) return;
     const foundItem = this.itemTargets.find(
       (item) => item.dataset.value === this.#cachedInputValue,
@@ -450,8 +418,8 @@ export default class Select2Controller extends Controller {
       }
       return;
     }
-    this.inputTarget.value = foundItem ? foundItem.dataset.label : "";
-    this.#updateAriaSelected(this.inputTarget.value);
+    this.triggerTarget.value = foundItem ? foundItem.dataset.label : "";
+    this.#updateAriaSelected(this.triggerTarget.value);
   }
 
   #updateAriaSelected(selectedLabel) {
@@ -465,11 +433,12 @@ export default class Select2Controller extends Controller {
 
   #updateAriaActiveDescendant() {
     const visibleItems = this.#visibleItems();
-    if (this.#currentItemIndex >= 0 && visibleItems[this.#currentItemIndex]) {
-      const activeId = visibleItems[this.#currentItemIndex].id;
-      this.inputTarget.setAttribute("aria-activedescendant", activeId);
+    const visibleItem = visibleItems[this.#currentItemIndex];
+    if (this.#currentItemIndex >= 0 && visibleItem) {
+      const activeId = visibleItem.id;
+      this.triggerTarget.setAttribute("aria-activedescendant", activeId);
     } else {
-      this.inputTarget.removeAttribute("aria-activedescendant");
+      this.triggerTarget.removeAttribute("aria-activedescendant");
     }
   }
 }
