+++ app/javascript/controllers/viral/dropdown_controller.js
@@ -1,29 +1,23 @@
-import MenuController from "controllers/menu_controller";
+import { Controller } from "@hotwired/stimulus";
 
-export default class extends MenuController {
+export default class extends Controller {
   static targets = ["trigger", "menu"];
+  static values = {
+    position: String,
+    trigger: String,
+    skidding: Number,
+    distance: Number,
+  };
 
   initialize() {
     this.boundOnButtonKeyDown = this.onButtonKeyDown.bind(this);
     this.boundOnButtonClick = this.onButtonClick.bind(this);
     this.boundOnMenuItemKeyDown = this.onMenuItemKeyDown.bind(this);
     this.boundFocusOut = this.focusOut.bind(this);
-    this.boundOnMorph = this.onMorph.bind(this);
   }
 
   connect() {
     this.element.setAttribute("data-controller-connected", "true");
-    document.addEventListener("turbo:morph", this.boundOnMorph);
-    this.#initializeDropdown();
-  }
-
-  disconnect() {
-    document.removeEventListener("turbo:morph", this.boundOnMorph);
-  }
-
-  onMorph() {
-    this.menuTargetConnected(this.menuTarget);
-    this.triggerTargetConnected(this.triggerTarget);
   }
 
   menuTargetConnected(element) {
@@ -45,36 +39,29 @@ export default class extends MenuController {
     element.addEventListener("click", this.boundOnButtonClick, {
       capture: true,
     });
-    super.triggerTargetConnected();
-  }
-
-  #initializeDropdown() {
-    super.share({
-      onShow: () => this.#onShow(),
-      onHide: () => this.#onHide(),
+    this.dropdown = new Dropdown(this.menuTarget, element, {
+      triggerType: "none",
+      offsetSkidding: this.skiddingValue,
+      offsetDistance: this.distanceValue,
+      onShow: () => {
+        this.triggerTarget.setAttribute("aria-expanded", "true");
+        this.menuTarget.setAttribute("aria-hidden", "false");
+        this.menuTarget.removeAttribute("hidden");
+      },
+      onHide: () => {
+        this.triggerTarget.setAttribute("aria-expanded", "false");
+        this.menuTarget.setAttribute("aria-hidden", "true");
+        this.menuTarget.setAttribute("hidden", "hidden");
+        this.#menuItems(element).forEach((menuitem) => {
+          menuitem.setAttribute("tabindex", "-1");
+        });
+      },
     });
-    super.connect();
-  }
-
-  #onShow() {
-    this.triggerTarget.setAttribute("aria-expanded", "true");
-    this.menuTarget.setAttribute("aria-hidden", "false");
-    this.menuTarget.removeAttribute("hidden");
-  }
-
-  #onHide() {
-    this.triggerTarget.setAttribute("aria-expanded", "false");
-    this.menuTarget.setAttribute("aria-hidden", "true");
-    this.menuTarget.setAttribute("hidden", "");
-    this.#menuItems(this.triggerTarget).forEach((menuitem) => {
-      menuitem.setAttribute("tabindex", "-1");
-    });
-    this.triggerTarget.focus();
   }
 
   focusOut(event) {
     if (!this.element.contains(event.relatedTarget)) {
-      super.hide();
+      this.dropdown.hide();
     }
   }
 
@@ -82,8 +69,8 @@ export default class extends MenuController {
     event.preventDefault();
     event.stopPropagation();
 
-    if (super.isVisible()) {
-      super.hide();
+    if (this.dropdown.isVisible()) {
+      this.dropdown.hide();
     } else {
       this.#openMenuAndFocusMenuItem(0);
     }
@@ -121,10 +108,9 @@ export default class extends MenuController {
         },
         { once: true },
       );
-      super.show();
-      this.#focusMenuItem(this.menuTarget);
+      this.dropdown.show();
     } else {
-      super.show();
+      this.dropdown.show();
       this.#focusMenuItem(menuItems.at(index));
     }
   }
@@ -149,7 +135,7 @@ export default class extends MenuController {
           if (clickableTarget) {
             clickableTarget.click();
           } else {
-            super.hide();
+            this.dropdown.hide();
           }
         } else {
           menuItems[currentIndex].click();
@@ -163,7 +149,7 @@ export default class extends MenuController {
         );
       case "Escape":
         event.preventDefault();
-        super.hide();
+        this.triggerTarget.focus();
         break;
       case "ArrowUp": {
         event.preventDefault();
@@ -199,7 +185,7 @@ export default class extends MenuController {
         if (event.shiftKey) {
           event.preventDefault();
           this.triggerTarget.focus();
-          super.hide();
+          this.dropdown.hide();
         }
         break;
     }
