+++ app/javascript/controllers/viral/select2_controller.js
@@ -1,4 +1,4 @@
-import MenuController from "controllers/menu_controller";
+import { Controller } from "@hotwired/stimulus";
 
 /**
  * Select2Controller
@@ -10,11 +10,11 @@ import MenuController from "controllers/menu_controller";
  * - Dropdown positioning and focus management
  * - Submit button enable/disable logic
  */
-export default class Select2Controller extends MenuController {
+export default class Select2Controller extends Controller {
   static targets = [
-    "trigger",
+    "input",
     "hidden",
-    "menu",
+    "dropdown",
     "scroller",
     "item",
     "empty",
@@ -26,6 +26,7 @@ export default class Select2Controller extends MenuController {
   #itemSelected = false;
   #cachedInputValue = "";
   #currentItemIndex = -1;
+  #dropdown = null;
   #boundHandlers = {};
 
   static #KEY_CODES = {
@@ -45,16 +46,16 @@ export default class Select2Controller extends MenuController {
 
       this.#boundHandlers.dropdownFocusOut =
         this.#handleDropdownFocusOut.bind(this);
-      this.menuTarget.addEventListener(
+      this.dropdownTarget.addEventListener(
         "focusout",
         this.#boundHandlers.dropdownFocusOut,
       );
 
       // Accessibility: set ARIA attributes
-      this.triggerTarget.setAttribute("role", "combobox");
-      this.triggerTarget.setAttribute("aria-autocomplete", "list");
-      this.triggerTarget.setAttribute("aria-expanded", "false");
-      this.triggerTarget.setAttribute(
+      this.inputTarget.setAttribute("role", "combobox");
+      this.inputTarget.setAttribute("aria-autocomplete", "list");
+      this.inputTarget.setAttribute("aria-expanded", "false");
+      this.inputTarget.setAttribute(
         "aria-controls",
         this.scrollerTarget.id || "select2-listbox",
       );
@@ -75,12 +76,15 @@ export default class Select2Controller extends MenuController {
   disconnect() {
     try {
       if (this.#boundHandlers.dropdownFocusOut) {
-        this.menuTarget.removeEventListener(
+        this.dropdownTarget.removeEventListener(
           "focusout",
           this.#boundHandlers.dropdownFocusOut,
         );
       }
-      super.hide();
+      if (this.#dropdown) {
+        this.#dropdown.hide();
+        this.#dropdown = null;
+      }
     } catch (error) {
       this.#handleError(error, "disconnect");
     }
@@ -115,8 +119,8 @@ export default class Select2Controller extends MenuController {
       if (selectedItemData) {
         const { value, label } = selectedItemData;
         this.#updateSelection(value, label);
-        super.hide();
-        this.triggerTarget.focus();
+        if (this.#dropdown) this.#dropdown.hide();
+        this.inputTarget.focus();
       } else {
         // If no valid item was determined (edge case or unexpected state),
         // potentially reset or log, but avoid throwing an error unless critical.
@@ -158,7 +162,7 @@ export default class Select2Controller extends MenuController {
           break;
         case Select2Controller.#KEY_CODES.ENTER:
           if (this.#currentItemIndex < 0) {
-            super.show();
+            if (this.#dropdown) this.#dropdown.show();
           } else {
             this.select(event);
           }
@@ -171,7 +175,8 @@ export default class Select2Controller extends MenuController {
 
   showDropdown() {
     try {
-      super.show();
+      if (this.#dropdown) this.#dropdown.show();
+      this.inputTarget.setAttribute("aria-expanded", "true");
     } catch (error) {
       this.#handleError(error, "showDropdown");
     }
@@ -179,7 +184,8 @@ export default class Select2Controller extends MenuController {
 
   hideDropdown() {
     try {
-      super.hide();
+      if (this.#dropdown) this.#dropdown.hide();
+      this.inputTarget.setAttribute("aria-expanded", "false");
     } catch (error) {
       this.#handleError(error, "hideDropdown");
     }
@@ -190,7 +196,7 @@ export default class Select2Controller extends MenuController {
    */
   input() {
     try {
-      const query = this.triggerTarget.value.toLowerCase().trim();
+      const query = this.inputTarget.value.toLowerCase().trim();
       let visibleItemCount = 0;
 
       this.#setItemSelected(false);
@@ -198,10 +204,10 @@ export default class Select2Controller extends MenuController {
       this.itemTargets.forEach((item) => {
         const text = item.textContent.toLowerCase() || "";
         if (text.includes(query)) {
-          item.classList.remove("hidden");
+          item.removeAttribute("hidden");
           visibleItemCount++;
         } else {
-          item.classList.add("hidden");
+          item.setAttribute("hidden", "");
         }
       });
 
@@ -209,8 +215,8 @@ export default class Select2Controller extends MenuController {
       this.#updateAriaActiveDescendant();
 
       if (visibleItemCount > 0) {
-        if (!super.isVisible()) {
-          super.show();
+        if (this.#dropdown && !this.#dropdown.isVisible()) {
+          this.#dropdown.show();
         }
         this.emptyTarget.setAttribute("hidden", "");
         this.scrollerTarget.scrollTop = 0;
@@ -241,8 +247,8 @@ export default class Select2Controller extends MenuController {
       );
       return; // Prevent updating with invalid data
     }
-    this.triggerTarget.value = label;
-    this.triggerTarget.title = label;
+    this.inputTarget.value = label;
+    this.inputTarget.title = label;
     this.#cachedInputValue = value; // Cache the *value*, not the label
     this.hiddenTarget.value = value;
     this.#updateAriaSelected(label); // Update ARIA state based on the new selection
@@ -259,17 +265,17 @@ export default class Select2Controller extends MenuController {
 
   #resetInput() {
     try {
-      super.hide();
+      if (this.#dropdown) this.#dropdown.hide();
       if (this.#cachedInputValue) {
         this.hiddenTarget.value = this.#cachedInputValue;
         this.#setInputTargetValueFromCache();
         this.#setItemSelected(true);
       } else {
-        this.triggerTarget.value = "";
+        this.inputTarget.value = "";
         this.hiddenTarget.value = "";
         this.#setItemSelected(false);
       }
-      this.triggerTarget.focus();
+      this.inputTarget.focus();
       this.#updateAriaActiveDescendant();
 
       if (this.hasSpreadsheetImportOutlet) {
@@ -331,32 +337,58 @@ export default class Select2Controller extends MenuController {
       // Check if the item itself or its immediate parent container is hidden
       const parentElement = item.closest("li") || item.parentNode; // Adjust selector if needed
       return (
-        !item.classList.contains("hidden") &&
-        !parentElement.classList.contains("hidden")
+        !item.hasAttribute("hidden") && !parentElement.hasAttribute("hidden")
       );
     });
   }
 
   #ensureItemVisible(item) {
     if (!this.scrollerTarget || !item) return;
-    item.scrollIntoView();
+    const container = this.scrollerTarget;
+    const containerRect = container.getBoundingClientRect();
+    const itemRect = item.getBoundingClientRect();
+    if (itemRect.bottom > containerRect.bottom) {
+      container.scrollTop += itemRect.bottom - containerRect.bottom;
+    } else if (itemRect.top < containerRect.top) {
+      container.scrollTop -= containerRect.top - itemRect.top;
+    }
   }
 
   #initializeDropdown() {
-    super.share({
-      onHide: () => this.#onHide(),
-    });
-    super.connect();
-  }
-
-  #onHide() {
-    if (!this.#itemSelected) this.#setInputTargetValueFromCache();
+    try {
+      if (typeof Dropdown !== "function") {
+        throw new Error(
+          "Flowbite Dropdown class not found. Make sure Flowbite JS is loaded.",
+        );
+      }
+      this.#dropdown = new Dropdown(this.dropdownTarget, this.inputTarget, {
+        placement: "bottom",
+        triggerType: "click",
+        offsetSkidding: 0,
+        offsetDistance: 10,
+        delay: 300,
+        onShow: () => {
+          this.dropdownTarget.style.minWidth = `${this.inputTarget.offsetWidth}px`;
+          this.inputTarget.setAttribute("aria-expanded", "true");
+          this.dropdownTarget.setAttribute("aria-hidden", "false");
+          this.dropdownTarget.removeAttribute("hidden");
+        },
+        onHide: () => {
+          this.inputTarget.setAttribute("aria-expanded", "false");
+          this.dropdownTarget.setAttribute("aria-hidden", "true");
+          this.dropdownTarget.setAttribute("hidden", "hidden");
+          if (!this.#itemSelected) this.#setInputTargetValueFromCache();
+        },
+      });
+    } catch (error) {
+      this.#handleError(error, "initializeDropdown");
+    }
   }
 
   #setDefaultSelection() {
     try {
-      if (!this.triggerTarget.value) return;
-      const query = this.triggerTarget.value;
+      if (!this.inputTarget.value) return;
+      const query = this.inputTarget.value;
       let matched = false;
       for (const item of this.itemTargets) {
         const { value, label } = item.dataset;
@@ -367,9 +399,9 @@ export default class Select2Controller extends MenuController {
           break;
         }
       }
-      if (!matched && this.triggerTarget.value) {
+      if (!matched && this.inputTarget.value) {
         throw new Error(
-          "No matching item found for the trigger value. Please check your data.",
+          "No matching item found for the input value. Please check your data.",
         );
       }
     } catch (error) {
@@ -379,8 +411,8 @@ export default class Select2Controller extends MenuController {
 
   #handleDropdownFocusOut(event) {
     try {
-      if (!this.menuTarget.contains(event.relatedTarget)) {
-        super.hide();
+      if (!this.dropdownTarget.contains(event.relatedTarget)) {
+        if (this.#dropdown) this.#dropdown.hide();
       }
     } catch (error) {
       this.#handleError(error, "handleDropdownFocusOut");
@@ -389,9 +421,9 @@ export default class Select2Controller extends MenuController {
 
   #validateTargets() {
     const missingTargets = [];
-    if (!this.hasTriggerTarget) missingTargets.push("trigger");
+    if (!this.hasInputTarget) missingTargets.push("input");
     if (!this.hasHiddenTarget) missingTargets.push("hidden");
-    if (!this.hasMenuTarget) missingTargets.push("menu");
+    if (!this.hasDropdownTarget) missingTargets.push("dropdown");
     if (!this.hasScrollerTarget) missingTargets.push("scroller");
     if (missingTargets.length > 0) {
       throw new Error(`Missing required targets: ${missingTargets.join(", ")}`);
@@ -404,7 +436,7 @@ export default class Select2Controller extends MenuController {
   }
 
   #setInputTargetValueFromCache() {
-    const inputValue = this.triggerTarget.value;
+    const inputValue = this.inputTarget.value;
     if (inputValue === this.#cachedInputValue) return;
     const foundItem = this.itemTargets.find(
       (item) => item.dataset.value === this.#cachedInputValue,
@@ -418,8 +450,8 @@ export default class Select2Controller extends MenuController {
       }
       return;
     }
-    this.triggerTarget.value = foundItem ? foundItem.dataset.label : "";
-    this.#updateAriaSelected(this.triggerTarget.value);
+    this.inputTarget.value = foundItem ? foundItem.dataset.label : "";
+    this.#updateAriaSelected(this.inputTarget.value);
   }
 
   #updateAriaSelected(selectedLabel) {
@@ -433,12 +465,11 @@ export default class Select2Controller extends MenuController {
 
   #updateAriaActiveDescendant() {
     const visibleItems = this.#visibleItems();
-    const visibleItem = visibleItems[this.#currentItemIndex];
-    if (this.#currentItemIndex >= 0 && visibleItem) {
-      const activeId = visibleItem.id;
-      this.triggerTarget.setAttribute("aria-activedescendant", activeId);
+    if (this.#currentItemIndex >= 0 && visibleItems[this.#currentItemIndex]) {
+      const activeId = visibleItems[this.#currentItemIndex].id;
+      this.inputTarget.setAttribute("aria-activedescendant", activeId);
     } else {
-      this.triggerTarget.removeAttribute("aria-activedescendant");
+      this.inputTarget.removeAttribute("aria-activedescendant");
     }
   }
 }
