+++ app/views/groups/_edit_advanced_transfer_form.html.erb
@@ -1,9 +1,9 @@
 <%= form_with(url: group_transfer_path, method: :put, class: "grid gap-4", id: "edit_advanced_transfer_form",
 data:
-if Flipper.enabled?(:flowbite_replacement)
-  { turbo_confirm: "", controller: "viral--select2", "viral--select2-trigger-type-value": "click" }
+if Flipper.enabled?(:beta_dropdown)
+  { turbo_confirm: "", controller: "viral--beta-select2", "viral--beta-select2-trigger-type-value": "click" }
 else
-  { turbo_confirm: "", controller: "viral--flowbite-select2", "viral--flowbite-select2-trigger-type-value": "click" }
+  { turbo_confirm: "", controller: "viral--select2", "viral--select2-trigger-type-value": "click" }
 end
 ) do |form| %>
   <%= render partial: "shared/form/required_field_legend" %>
@@ -61,15 +61,15 @@ end
                 class: "button button-destructive inline-block",
                 disabled: true,
                 data:
-                  if Flipper.enabled?(:flowbite_replacement)
+                  if Flipper.enabled?(:beta_dropdown)
                     {
-                      "viral--select2-target": "submitButton",
+                      "viral--beta-select2-target": "submitButton",
                       turbo_content: ".confirm-transfer",
                       confirm_value: confirm_value,
                     }
                   else
                     {
-                      "viral--flowbite-select2-target": "submitButton",
+                      "viral--select2-target": "submitButton",
                       turbo_content: ".confirm-transfer",
                       confirm_value: confirm_value,
                     }
