+++ app/views/groups/group_links/_form.html.erb
@@ -8,10 +8,10 @@
 </div>
 
 <%= form_with(model: new_group_link, url: group_group_links_path(@namespace, tab: @tab), data:
-if Flipper.enabled?(:flowbite_replacement)
-  { controller: "viral--select2", "viral--select2-trigger-type-value": "click" }
+if Flipper.enabled?(:beta_dropdown)
+  { controller: "viral--beta-select2", "viral--beta-select2-trigger-type-value": "click" }
 else
-  { controller: "viral--flowbite-select2", "viral--flowbite-select2-trigger-type-value": "click" }
+  { controller: "viral--select2", "viral--select2-trigger-type-value": "click" }
 end, html: { novalidate: true }) do |form| %>
   <div class="grid gap-4">
     <%= if new_group_link.errors.any?
@@ -117,10 +117,10 @@ end, html: { novalidate: true }) do |form| %>
                   class: "button button-primary",
                   disabled: true,
                   data:
-                    if Flipper.enabled?(:flowbite_replacement)
-                      { "viral--select2-target": "submitButton" }
+                    if Flipper.enabled?(:beta_dropdown)
+                      { "viral--beta-select2-target": "submitButton" }
                     else
-                      { "viral--flowbite-select2-target": "submitButton" }
+                      { "viral--select2-target": "submitButton" }
                     end %>
       <button
         type="button"
