+++ app/views/groups/new_subgroup.html.erb
@@ -5,10 +5,10 @@
 
 <div data-controller="slugify" class="@xl:container">
   <%= form_with(model: @new_group, url: groups_path, method: :post, data:
-  if Flipper.enabled?(:flowbite_replacement)
-  { controller: "viral--select2", "viral--select2-trigger-type-value": "click" }
+  if Flipper.enabled?(:beta_dropdown)
+  { controller: "viral--beta-select2", "viral--beta-select2-trigger-type-value": "click" }
 else
-  { controller: "viral--flowbite-select2", "viral--flowbite-select2-trigger-type-value": "click" }
+  { controller: "viral--select2", "viral--select2-trigger-type-value": "click" }
 end,
    html: { novalidate: true }) do |form| %>
     <div class="grid gap-4">
