+++ app/views/shared/samples/_metadata_template_dropdown.html.erb
@@ -1,18 +1,18 @@
 <div
   class="flex max-sm:grow"
-  <% if Flipper.enabled?(:flowbite_replacement) %>
-  data-controller="viral--dropdown"
+  <% if Flipper.enabled?(:beta_dropdown) %>
+  data-controller="viral--beta-dropdown"
   <% else %>
-  data-controller="viral--flowbite-dropdown"
-  data-viral--flowbite-dropdown-skidding-value="0"
-  data-viral--flowbite-dropdown-distance-value="10"
+  data-controller="viral--dropdown"
+  data-viral--dropdown-skidding-value="0"
+  data-viral--dropdown-distance-value="10"
   <% end %>
 >
   <button
-    <% if Flipper.enabled?(:flowbite_replacement) %>
-    data-viral--dropdown-target="trigger"
+    <% if Flipper.enabled?(:beta_dropdown) %>
+    data-viral--beta-dropdown-target="trigger"
     <% else %>
-    data-viral--flowbite-dropdown-target="trigger"
+    data-viral--dropdown-target="trigger"
     <% end %>
     title="<%= t("shared.samples.metadata_templates.tooltip") %>"
     id="metadata-template-dd-trigger"
@@ -29,10 +29,10 @@
   </button>
   <!-- Dropdown menu -->
   <% options = { hidden: "" } %>
-  <% if Flipper.enabled?(:flowbite_replacement) %>
-    <% options["data-viral--dropdown-target"] = "menu" %>
+  <% if Flipper.enabled?(:beta_dropdown) %>
+    <% options["data-viral--beta-dropdown-target"] = "menu" %>
   <% else %>
-    <% options["data-viral--flowbite-dropdown-target"] = "menu" %>
+    <% options["data-viral--dropdown-target"] = "menu" %>
   <% end %>
   <%= turbo_frame_tag "metadata-template-dd-menu", class: "
        w-48 bg-white divide-y divide-slate-100 rounded-lg shadow-sm
