+++ embedded_gems/pathogen/app/assets/javascripts/pathogen_view_components.js
@@ -2,7 +2,9 @@
 import TabsController from "pathogen_view_components/tabs_controller";
 import TooltipController from "pathogen_view_components/tooltip_controller";
 import DatepickerInputController from "pathogen_view_components/datepicker/input_controller";
+import DatepickerBetaInputController from "pathogen_view_components/datepicker/beta_input_controller";
 import DatepickerCalendarController from "pathogen_view_components/datepicker/calendar_controller";
+import DatepickerBetaCalendarController from "pathogen_view_components/datepicker/beta_calendar_controller";
 
 function registerPathogenControllers(application) {
   if (!application || typeof application.register !== "function") {
@@ -17,8 +19,16 @@ function registerPathogenControllers(application) {
     DatepickerInputController,
   );
   application.register(
+    "pathogen--datepicker--beta-input",
+    DatepickerBetaInputController,
+  );
+  application.register(
     "pathogen--datepicker--calendar",
     DatepickerCalendarController,
+  );
+  application.register(
+    "pathogen--datepicker--beta-calendar",
+    DatepickerBetaCalendarController,
   );
 
   if (import.meta.env?.DEV) {
@@ -28,7 +38,9 @@ function registerPathogenControllers(application) {
 
 export {
   DatepickerCalendarController,
+  DatepickerBetaCalendarController,
   DatepickerInputController,
+  DatepickerBetaInputController,
   TabsController,
   TooltipController,
   registerPathogenControllers,
