+++ embedded_gems/pathogen/app/assets/javascripts/pathogen_view_components/datepicker/beta_input_controller.js
