+++ embedded_gems/pathogen/app/components/pathogen/datepicker/_input_field.html.erb
@@ -13,17 +13,17 @@
     placeholder="<%= t("pathogen.datepicker.placeholder") %>"
     autocomplete="off"
     value="<%= selected_date %>"
-    <% if Flipper.enabled?(:flowbite_replacement) %>
+    <% if Flipper.enabled?(:beta_dropdown) %>
+    data-pathogen--datepicker--beta-input-target="trigger"
+    data-action="
+      change->pathogen--datepicker--beta-input#directInput
+      keydown->pathogen--datepicker--beta-input#handleEnterDirectInput
+    "
+    <% else %>
     data-pathogen--datepicker--input-target="trigger"
     data-action="
       change->pathogen--datepicker--input#directInput
       keydown->pathogen--datepicker--input#handleEnterDirectInput
-    "
-    <% else %>
-    data-pathogen--datepicker--flowbite-input-target="trigger"
-    data-action="
-      change->pathogen--datepicker--flowbite-input#directInput
-      keydown->pathogen--datepicker--flowbite-input#handleEnterDirectInput
     "
     <% end %>
     type="text"
@@ -40,10 +40,10 @@
 </div>
 
 <div
-  <% if Flipper.enabled?(:flowbite_replacement) %>
-  data-pathogen--datepicker--input-target="inputError"
+  <% if Flipper.enabled?(:beta_dropdown) %>
+  data-pathogen--datepicker--beta-input-target="inputError"
   <% else %>
-  data-pathogen--datepicker--flowbite-input-target="inputError"
+  data-pathogen--datepicker--input-target="inputError"
   <% end %>
   class="mt-1 hidden text-sm text-red-700 dark:text-red-500"
   aria-hidden="true"
