+++ embedded_gems/pathogen/app/components/pathogen/datepicker/calendar/_header.html.erb
@@ -15,45 +15,41 @@
         enabled:dark:hover:text-white text-lg p-2.5 prev-btn mr-2 enabled:cursor-pointer
         disabled:cursor-not-allowed
       "
-      <% if Flipper.enabled?(:flowbite_replacement) %>
+      <% if Flipper.enabled?(:beta_dropdown) %>
+      data-action="
+        click->pathogen--datepicker--beta-calendar#previousMonth
+        keydown->pathogen--datepicker--beta-calendar#tabBackToInput
+      "
+      data-pathogen--datepicker--beta-calendar-target="backButton"
+      <% else %>
       data-action="
         click->pathogen--datepicker--calendar#previousMonth
         keydown->pathogen--datepicker--calendar#tabBackToInput
       "
       data-pathogen--datepicker--calendar-target="backButton"
-      <% else %>
-      data-action="
-        click->pathogen--datepicker--flowbite-calendar#previousMonth
-        keydown->pathogen--datepicker--flowbite-calendar#tabBackToInput
-      "
-      data-pathogen--datepicker--flowbite-calendar-target="backButton"
       <% end %>
       aria-label="<%= t("pathogen.datepicker.aria_label.previous_month_button")%>"
     >
       <%= pathogen_icon(:arrow_left, size: :sm) %>
     </button>
     <div
-      <% if Flipper.enabled?(:flowbite_replacement) %>
+      <% if Flipper.enabled?(:beta_dropdown) %>
+      data-pathogen--datepicker--beta-calendar-target="monthSelectContainer"
+      data-action="keydown->pathogen--datepicker--beta-calendar#tabBackToInput"
+      <% else %>
       data-pathogen--datepicker--calendar-target="monthSelectContainer"
       data-action="keydown->pathogen--datepicker--calendar#tabBackToInput"
-      <% else %>
-      data-pathogen--datepicker--flowbite-calendar-target="monthSelectContainer"
-      data-action="
-        keydown->pathogen--datepicker--flowbite-calendar#tabBackToInput
-      "
       <% end %>
     >
     </div>
     <div>
       <input
-        <% if Flipper.enabled?(:flowbite_replacement) %>
+        <% if Flipper.enabled?(:beta_dropdown) %>
+        data-pathogen--datepicker--beta-calendar-target="year"
+        data-action="change->pathogen--datepicker--beta-calendar#changeYear"
+        <% else %>
         data-pathogen--datepicker--calendar-target="year"
         data-action="change->pathogen--datepicker--calendar#changeYear"
-        <% else %>
-        data-pathogen--datepicker--flowbite-calendar-target="year"
-        data-action="
-          change->pathogen--datepicker--flowbite-calendar#changeYear
-        "
         <% end %>
         type="number"
         max="9999"
@@ -72,10 +68,10 @@
         hover:bg-slate-100 dark:hover:bg-slate-600 hover:text-slate-900
         dark:hover:text-white text-lg p-2.5 next-btn ml-2 cursor-pointer
       "
-      <% if Flipper.enabled?(:flowbite_replacement) %>
-      data-action="click->pathogen--datepicker--calendar#nextMonth"
+      <% if Flipper.enabled?(:beta_dropdown) %>
+      data-action="click->pathogen--datepicker--beta-calendar#nextMonth"
       <% else %>
-      data-action="click->pathogen--datepicker--flowbite-calendar#nextMonth"
+      data-action="click->pathogen--datepicker--calendar#nextMonth"
       <% end %>
       aria-label="<%= t("pathogen.datepicker.aria_label.next_month_button")%>"
     >
@@ -85,19 +81,19 @@
 </div>
 
 <template
-  <% if Flipper.enabled?(:flowbite_replacement) %>
-  data-pathogen--datepicker--calendar-target="monthSelectTemplate"
+  <% if Flipper.enabled?(:beta_dropdown) %>
+  data-pathogen--datepicker--beta-calendar-target="monthSelectTemplate"
   <% else %>
-  data-pathogen--datepicker--flowbite-calendar-target="monthSelectTemplate"
+  data-pathogen--datepicker--calendar-target="monthSelectTemplate"
   <% end %>
 >
   <select
-    <% if Flipper.enabled?(:flowbite_replacement) %>
+    <% if Flipper.enabled?(:beta_dropdown) %>
+    data-pathogen--datepicker--beta-calendar-target="monthSelect"
+    data-action="change->pathogen--datepicker--beta-calendar#changeMonth"
+    <% else %>
     data-pathogen--datepicker--calendar-target="monthSelect"
     data-action="change->pathogen--datepicker--calendar#changeMonth"
-    <% else %>
-    data-pathogen--datepicker--flowbite-calendar-target="monthSelect"
-    data-action="change->pathogen--datepicker--flowbite-calendar#changeMonth"
     <% end %>
     class="
       rounded-lg border-slate-300 text-slate-900 sm:text-sm dark:border-slate-600
