+++ test/components/previews/viral_select2_component_preview/default.html.erb
@@ -1,8 +1,8 @@
 <%= form_with(url: "#", data:
-                      if Flipper.enabled?(:flowbite_replacement)
-                        { controller: "viral--select2", "viral--select2-trigger-type-value": "click" }
+                      if Flipper.enabled?(:beta_dropdown)
+                        { controller: "viral--beta-select2", "viral--beta-select2-trigger-type-value": "click" }
                       else
-                        { controller: "viral--flowbite-select2", "viral--flowbite-select2-trigger-type-value": "click" }
+                        { controller: "viral--select2", "viral--select2-trigger-type-value": "click" }
                       end
   ) do |form| %>
   <label for="unique_id">Select User</label>
@@ -29,9 +29,9 @@
               class: "button button-primary",
               disabled: true,
               data:
-                if Flipper.enabled?(:flowbite_replacement)
-                  { "viral--select2-target": "submitButton" }
+                if Flipper.enabled?(:beta_dropdown)
+                  { "viral--beta-select2-target": "submitButton" }
                 else
-                  { "viral--flowbite-select2-target": "submitButton" }
+                  { "viral--select2-target": "submitButton" }
                 end %>
 <% end %>
