+++ test/components/viral/select2_component_test.rb
@@ -3,16 +3,13 @@
 require 'application_system_test_case'
 
 class Select2ComponentTest < ApplicationSystemTestCase
-  def setup
-    Flipper.enable(:flowbite_replacement)
-  end
   test 'default' do
     visit '/rails/view_components/viral_select2_component/default'
     assert_selector 'input[type="hidden"][name="user"]', visible: :hidden, count: 1
     assert_selector 'input[type="submit"][disabled]', count: 1
 
     find('input.select2-input[type="text"]').click
-    assert_selector 'div[data-viral--select2-target="menu"]', visible: :visible
+    assert_selector 'div[data-viral--select2-target="dropdown"]', visible: :visible
     assert_selector 'ul[data-viral--select2-target="scroller"] li', count: 50
 
     find('li[data-label="User 1"]').click
