+++ test/system/groups/group_links_test.rb
@@ -13,7 +13,6 @@ module Groups
       @group_links_count = namespace_group_links.select { |group_link| group_link.namespace == @namespace }.count
       @group_link5 = namespace_group_links(:namespace_group_link5)
       @group_link14 = namespace_group_links(:namespace_group_link14)
-      Flipper.enable(:flowbite_replacement)
     end
 
     test 'can create a group to group link' do
@@ -32,7 +31,7 @@ module Groups
           :'groups.group_links.new.sharing_namespace_with_group',
           name: @namespace.human_name
         )
-        find('input[data-viral--select2-target="trigger"]').click
+        find('input[data-viral--select2-target="input"]').click
         find("li[data-label='Subgroup 3']").click
         find('#namespace_group_link_group_access_level').find(:xpath, 'option[3]').select_option
