+++ test/system/groups_test.rb
@@ -16,7 +16,6 @@ class GroupsTest < ApplicationSystemTestCase
     @project30 = projects(:project30)
     @sample34 = samples(:sample34)
     login_as @user
-    Flipper.enable(:flowbite_replacement)
   end
 
   test 'can create a group' do
@@ -127,7 +126,7 @@ class GroupsTest < ApplicationSystemTestCase
     click_link I18n.t('groups.show.create_subgroup_button')
 
     within %(div[data-controller="slugify"][data-controller-connected="true"]) do
-      assert_selector %(input[data-viral--select2-target="trigger"]) do |input|
+      assert_selector %(input[data-viral--select2-target="input"]) do |input|
         assert_equal group1.path, input['value']
       end
     end
